Added modules and component for DI Also the Room has been added (with simple repository)

+++ app/src/main/java/com/tommy/simpleweather/storage/dao/UserDAO.kt
@@ -8,5 +8,5 @@ interface UserDAO {
 
     @Transaction
     @Query("SELECT * FROM user WHERE userId = :id")
-    fun getById(id: Int): UserWithCitiesAndForecast?
+    fun getById(id: Long): UserWithCitiesAndForecast?
 }
